refactor: refactor code indent
