+++ src/main/java/racingcar/domain/Car.java
@@ -39,5 +39,4 @@ public class Car {
     public int getPosition() {
         return this.position;
     }
-
 }
