+++ src/main/java/racingcar/view/InputView.java
@@ -27,5 +27,4 @@ public class InputView {
         System.out.println("시도할 횟수는 몇 회인가요?");
         return sc.nextLine();
     }
-
 }
